Create ItemsManager, ItemTier, and ItemResource.
ItemsManager: Is a scene holding a array of items and tiers of items. Provides functions for getting a random item at a certain tier or just in general.

ItemTier: Array of items all at a specific tier. An array of these is generated at runtime by ItemsManager

ItemResource: Reference to an item scene.