Initital weapons and items (#6)
* Add Item, Ammo, AmmoResource, and WEapon

Abstract classes for items.

* Add MeleWeapon and RangedWeapon

MeleWeapon and RangedWeapon split the functionality of Weapon into one that reloads, uses ammo, and shoots the attacker, and one that doesn't do those things and has a hitbox built in.

* Add Fists, player can now attack.

Player can now attack with the mouse, what do we think?

Also created a new timer resource that is better.

* Make custom health, hitbox, hurtbox, and implement windup

* Fix input just_pressed clearing on every input.

Resolves #7

* Finishing up, ready for testing